refactor: rewrite ls command into list.js
